requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-imagekit==4.1.0
idna==3.4
oauthlib==3.2.2
pilkit==2.0
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
